% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{clean_education}
\alias{clean_education}
\title{Cleaning screener education responses}
\usage{
clean_education(data)
}
\arguments{
\item{data}{data frame to act upon}
}
\value{
a dataframe with the education responses cleaned
}
\description{
This function cleans education responses
}
