% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_all_data}
\alias{get_all_data}
\title{Pull ORCA Data}
\usage{
get_all_data(token)
}
\arguments{
\item{token}{Unique REDCap API token}
}
\value{
A data frame with all records, survey responses & events
}
\description{
pulls all records & survey responses for ORCA 1.0
}
