% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_orca_bitsea}
\alias{get_orca_bitsea}
\title{Process 24m BITSEA Survey Data}
\usage{
get_orca_bitsea(token, timestamp = T)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timestamp}{Boolean for whether to include survey timestamp}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the sum scores for the MCHAT subscales and threshold information.
}
