% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_orca_demo}
\alias{get_orca_demo}
\title{Pulls sociodemographic cleaned responses}
\usage{
get_orca_demo(token)
}
\arguments{
\item{token}{The API token for the project}
}
\value{
the completed sociodemographic responses with numeric categories replaced with character
}
\description{
This cleans and pulls demographic information
}
