% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_orca_dq}
\alias{get_orca_dq}
\title{Pulling ORCA Data quality}
\usage{
get_orca_dq(token, tasks = c("cecile", "maap", "memory", "vpc"))
}
\arguments{
\item{token}{The API token for the project}

\item{tasks}{The tasks you wish to export data quality for - if multiple, place in vector}
}
\value{
A data frame with the retrieved data
}
\description{
RETURNS DATA QUALITY SCORES FOR EACH TASK SPECIFIED
}
