% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_orca_ecbq}
\alias{get_orca_ecbq}
\title{Process 24m ECBQ Survey Data}
\usage{
get_orca_ecbq(token, timestamp = T)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timestamp}{Boolean for whether to include survey timestamp}
}
\value{
A data frame for the completed surveys for 24m
}
\description{
This function will download and return the mean scores for the ECBQ subscales.
}
