% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_orca_epds}
\alias{get_orca_epds}
\title{Process EPDS postpartum depression data}
\usage{
get_orca_epds(token, timepoint = "orca_visit_1_arm_1", timestamp = T)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{Redcap event name for timepoint}

\item{timestamp}{Boolean - whether to include survey timestamps}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the totalsummed scores for the EPDS. Only total scores, and not cutoff values, are returned.
}
