% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_orca_family_routines}
\alias{get_orca_family_routines}
\title{Process Family Routines Data}
\usage{
get_orca_family_routines(
  token,
  timepoint = "orca_visit_1_arm_1",
  timestamp = T
)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{redcap event name for timepoint}

\item{timestamp}{Boolean for whether to include survey timestamp}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return relevant fields from family routines, filtered by timepoint
}
