% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_orca_mchat}
\alias{get_orca_mchat}
\title{Process 24m MCHAT Survey Data}
\usage{
get_orca_mchat(token, timestamp = T)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timestamp}{Boolean for whether to include survey timestamp}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the sum score for the MCHAT.
}
