% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_orca_screener}
\alias{get_orca_screener}
\title{Pulling raw orca screener form}
\usage{
get_orca_screener(token = token, form = form, raw_v_label = "raw")
}
\arguments{
\item{token}{REDCap API token for ORCA screener redcap project}

\item{form}{name of form from the ORCA screener redcap project}

\item{raw_v_label}{whether to pull raw or labelled responses}
}
\value{
a dataframe of the form pulled
}
\description{
This function pulls raw forms from ORCA screener redcap project
}
