% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrcaData-package.R
\docType{package}
\name{OrcaData-package}
\alias{OrcaData}
\alias{OrcaData-package}
\title{OrcaData: Pulls data from ORCA redcap projects via API}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Amy Hume \email{amyhume66@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
