% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_functions.r
\name{get_orca_md}
\alias{get_orca_md}
\title{Process Material Deprivation data}
\usage{
get_orca_md(token, timestamp = T)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timestamp}{Boolean - whether to include survey timestamps}
}
\value{
A data frame with the calculated md score for each ID
}
\description{
This function will download and return the total summed scores
}
